+++ apps/storybook/src/components/layout/baseLayout.tsx
@@ -1,14 +1,14 @@
 import { FC } from "react";
-import { Scaffold } from "@orderly.network/ui-scaffold";
+import { Scaffold, ScaffoldProps } from "@orderly.network/ui-scaffold";
 import { useNav } from "../../hooks/useNav";
 import { useOrderlyConfig } from "../../hooks/useOrderlyConfig";
 
-type OrderlyLayoutProps = {
+type BaseLayoutProps = {
   children: React.ReactNode;
   initialMenu?: string;
-};
+} & Pick<ScaffoldProps, "classNames">;
 
-export const OrderlyLayout: FC<OrderlyLayoutProps> = (props) => {
+export const BaseLayout: FC<BaseLayoutProps> = (props) => {
   const { onRouteChange } = useNav();
   const config = useOrderlyConfig();
 
@@ -22,6 +22,7 @@ export const OrderlyLayout: FC<OrderlyLayoutProps> = (props) => {
       routerAdapter={{
         onRouteChange,
       }}
+      classNames={props.classNames}
     >
       {props.children}
     </Scaffold>