+++ apps/storybook/src/stories/package/portfolio/positions.stories.tsx
@@ -5,7 +5,7 @@ import {
 } from "@orderly.network/portfolio";
 import { useTradingLocalStorage } from "@orderly.network/trading";
 import { Box } from "@orderly.network/ui";
-import { PortfolioLayout } from "../../../components/layout/portfolioLayout";
+import { PortfolioLayout } from "../../../components/layout";
 import { useOrderlyConfig } from "../../../hooks/useOrderlyConfig";
 
 const meta: Meta<typeof PositionsModule.PositionsPage> = {