+++ packages/ui-scaffold/src/components/restrictedInfo/restrictedInfo.script.tsx
@@ -1,6 +1,7 @@
+import { useRef, useState } from "react";
 import { useAppContext } from "@orderly.network/react-app";
 import { useConfig } from "@orderly.network/hooks";
-import { useEffect, useRef, useState } from "react";
+import { useObserverElement } from "@orderly.network/ui";
 
 export type UseRestrictedInfoScriptReturn = ReturnType<
   typeof useRestrictedInfoScript
@@ -14,24 +15,9 @@ export const useRestrictedInfoScript = () => {
   const container = useRef<HTMLDivElement>(null);
   const [mutiLine, setMutiLine] = useState(false);
 
-  useEffect(() => {
-    const element = container.current;
-
-    if (!element) return;
-
-    const resizeObserver = new ResizeObserver((entries) => {
-      for (let entry of entries) {
-        const height = entry.contentRect.height;
-        setMutiLine(height > 28);
-      }
-    });
-
-    resizeObserver.observe(element);
-
-    return () => {
-      resizeObserver.unobserve(element);
-    };
-  }, []);
+  useObserverElement(container.current, (entry) => {
+    setMutiLine(entry.contentRect.height > 28);
+  });
 
   return {
     restrictedInfo,