+++ packages/ui/src/hooks/index.ts
@@ -1,1 +1,3 @@
 export { useScreen } from "./useScreen";
+export { useMediaQuery } from "./useMediaQuery";
+export { useObserverElement } from "./useObserverElement";