+++ packages/create-orderly-app/CHANGELOG.md
@@ -1,5 +1,21 @@
 # create-orderly-app
 
+## 1.6.0
+
+### Minor Changes
+
+- publish
+
+### Patch Changes
+
+- ba163f4ce: publish
+
+## 1.5.24-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 1.5.23
 
 ### Patch Changes