+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,21 @@
 # tsconfig
 
+## 0.4.0
+
+### Minor Changes
+
+- publish
+
+### Patch Changes
+
+- ba163f4ce: publish
+
+## 0.3.23-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 0.3.22
 
 ### Patch Changes