+++ packages/wallet-connector-privy/CHANGELOG.md
@@ -1,1 +1,34 @@
 # @orderly.network/wallet-connector-privy
+
+## 2.1.0
+
+### Minor Changes
+
+- publish
+
+### Patch Changes
+
+- ba163f4ce: publish
+- Updated dependencies [ba163f4ce]
+- Updated dependencies
+  - @orderly.network/core@2.1.0
+  - @orderly.network/hooks@2.1.0
+  - @orderly.network/i18n@2.1.0
+  - @orderly.network/types@2.1.0
+  - @orderly.network/ui@2.1.0
+  - @orderly.network/ui-connector@2.1.0
+  - @orderly.network/utils@2.1.0
+
+## 2.0.8-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/core@2.0.8-alpha.0
+  - @orderly.network/hooks@2.0.8-alpha.0
+  - @orderly.network/i18n@2.0.8-alpha.0
+  - @orderly.network/types@2.0.8-alpha.0
+  - @orderly.network/ui@2.0.8-alpha.0
+  - @orderly.network/ui-connector@2.0.8-alpha.0
+  - @orderly.network/utils@2.0.8-alpha.0